feat(website): add responsive Hamburger menu for small screens
resolves #220
