+++ website/src/layouts/base/header/HamburgerMenu.astro
@@ -1,0 +1,62 @@
+---
+import { getSession } from 'auth-astro/server';
+
+import HamburgerMenuItem from './HamburgerMenuItem.astro';
+import HamburgerMenuSection from './HamburgerMenuSection.astro';
+import { getPathogenMegaMenuSections } from './getPathogenMegaMenuSections';
+import { LoginButton } from '../../../components/auth/LoginButton';
+import { Page } from '../../../types/pages';
+
+type Props = {
+    forceLoggedOutState: boolean;
+};
+
+const { forceLoggedOutState } = Astro.props;
+
+const pathogenMegaMenuSections = getPathogenMegaMenuSections();
+
+const session = await getSession(Astro.request);
+const showLoggedInState = !forceLoggedOutState && session?.user !== undefined;
+---
+
+<nav>
+    <input id='menu__toggle' class='peer hidden' type='checkbox' />
+    <label
+        for='menu__toggle'
+        class='iconify absolute right-5 top-4 cursor-pointer text-4xl mdi--hamburger-menu peer-checked:mdi--window-close'
+    >
+    </label>
+
+    <ul
+        class='fixed left-full top-0 mt-16 h-[calc(100dvh-4rem)] w-full overflow-scroll border bg-white pl-4 duration-200 peer-checked:left-0'
+    >
+        {
+            pathogenMegaMenuSections.map((section) => (
+                <HamburgerMenuSection navigationEntries={section.navigationEntries}>
+                    {section.headline}
+                </HamburgerMenuSection>
+            ))
+        }
+        <HamburgerMenuItem href={Page.dataSources} itemType='toplevel'>Data Sources </HamburgerMenuItem>
+        {
+            showLoggedInState ? (
+                <HamburgerMenuSection
+                    navigationEntries={[
+                        { label: session.user!.name! },
+                        {
+                            label: 'Subscriptions',
+                            href: '/subscriptions',
+                        },
+                        { label: 'Logout', href: '/logout' },
+                    ]}
+                >
+                    My Account
+                </HamburgerMenuSection>
+            ) : (
+                <HamburgerMenuItem itemType='toplevel'>
+                    <LoginButton client:load />
+                </HamburgerMenuItem>
+            )
+        }
+    </ul>
+</nav>
